requirements.txt
PyYAML
click
pyobjc
numpy
